Exports `glm::ext::Consts`.
Exports trait `glm::ext::Consts` so that types of scalar constants can
be specified by calling methods in this trait.

For example, `let pi2 = pi() * 2_f32;` can not be compiled because
return type of generic function `pi()` could be scalar or vectors. If a
scalar is expected, the statement can be written as `let pi2 =
f32::pi() * 2.;`.